imobiliaria_api/requirements.txt
alembic==1.15.1
aniso8601==10.0.0
cffi==1.17.1
click==8.1.8
cryptography==41.0.1
Flask==2.2.3
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
gunicorn==20.1.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==24.2
pluggy==1.5.0
pycparser==2.22
PyMySQL==1.0.3
pytest==7.3.1
python-dotenv==1.0.0
pytz==2025.1
setuptools==77.0.1
six==1.17.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
